requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.28
openai>=1.40.0
pytest==8.1.1
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
